+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 **Improvements and fixes**
 - Main screen - Improvements to some cut off texts
+- Main screen - Improve the “two blocks per row” threshold when using custom font scale
 - Details - Fix precipitation probability details being expressed in precipitation unit instead of %
 - Fix missing normals every other refresh